Integrate systemd networkd - copy files from debian-rescue-initramfs repo
Migrate the network configuration to use the systemd services.
This improves the speed that networks are configured and also brings
DHCP servers, LLDP services and NAT sharing of networks.

This change required complex configuration to be done on the debian root
image.  The simplest way to do this was to copy an existing framework
from another project.  The files added here have come from an older
debian ramdisk builder (with some minor changes to the README file).

It is my intent to keep these files in sync between the two projects
(curently by manually copying them - but time will tell if that is
scalable)

Originally publicly published at
    https://github.com/hamishcoleman/debian-rescue-initramfs
copied from
    commit d450c3cf820737658ba8ddae50e6b93b24ce6ff7