The u-boot board name was changed in the newer version !!!
I'm not sure why you would change this and it is unfortunate, since
I have used that as a key value to make the name of the output file.

For the moment, I am adding some hacks to fudge the name of the DTB
to be the one that u-boot is now expecting.

In the future, I hope that we will be able to update the kernel
version once the xradio_wifi support is working again and remove
this hack.

At that point, all the documentation about what output file name to
expect will need to change.