BUG: fix build on newer versions of make.
The old version of make was able to infer an extra dependancy somewhere
(or perhaps evaulated the source tree contents at different points)

The new version of make does not see the source for the dtbo in existance
when it is first run, so never tries to use the builder line.  The source
only turns up after the armbian package is unpacked, so a second try
at making will work - however, just assuming that the armbian tag will
generate the right source files for us is more reliable.