Work around strange issue with booting on banana pi.
When booting with default config, u-boot had failures while trying to
load the partition table:

    mmc0 is current device
    bad MBR sector signature 0x088b
    bad MBR sector signature 0x288b
    ** Invalid partition 1 **

After which it would give up on mmc and move onto the next boot type.
The u-boot source code shows that the signature it is looking for is
the usual 0x55 0xaa one at the end of a valid boot block.  The line in
the embedded u-boot config that caused the error was:

    mmc_boot=if mmc dev ${devnum}; then setenv devtype mmc; run scan_dev_for_boot_part; fi

The "if mmc dev ${devnum}" command appeared to be the failing part - even
though u-boot would happily use the partition table.

Changing the bootcmd to:

    bootcmd=run loadsplash; setenv devnum 0; setenv devtype mmc; run scan_dev_for_boot_part

bypasses the normal "run distro_bootcmd" and forcing mmc boot, allowing it
to boot fine.

FIXME - why is u-boot claiming that there is a corrupted MBR?  Is there an
issue with the mmc driver?